Tutorial with gypsum solubility is replaced by the anhydrite solubility
